move segment top k collector
