+++ quickwit/quickwit-search/src/lib.rs
@@ -42,6 +42,7 @@ mod search_response_rest;
 mod search_stream;
 mod service;
 mod thread_pool;
+pub(crate) mod top_k_collector;
 
 mod metrics;
